refactor: UseCaseに切り出した
