+++ modules/model/src/main/java/net/pantasystem/milktea/model/user/follow/requests/RejectFollowRequestUseCase.kt
@@ -1,0 +1,15 @@
+package net.pantasystem.milktea.model.user.follow.requests
+
+import net.pantasystem.milktea.common.runCancellableCatching
+import net.pantasystem.milktea.model.UseCase
+import net.pantasystem.milktea.model.user.FollowRequestRepository
+import net.pantasystem.milktea.model.user.User
+import javax.inject.Inject
+
+class RejectFollowRequestUseCase @Inject constructor(
+    private val followRequestRepository: FollowRequestRepository,
+): UseCase {
+    suspend operator fun invoke(userId: User.Id) = runCancellableCatching {
+        followRequestRepository.reject(userId)
+    }
+}
